Extract subprocess tests to own file
