+++ tests/e2e/specs/Tasks.spec.js
@@ -1,6 +1,5 @@
 import {
   addNodeTypeToPaper,
-  assertDownloadedXmlContainsExpected,
   dragFromSourceToDest,
   getElementAtPosition,
   modalCancel,
@@ -33,41 +32,6 @@ describe('Tasks', () => {
     waitToRenderAllShapes();
 
     getElementAtPosition(taskPosition).getType().should('equal', nodeTypes.task);
-  });
-
-  it('Can create sub process with config', () => {
-    const processName = 'Process with multiple start events';
-    const subProcessPosition = { x: 250, y: 250 };
-    const subProcessName = 'Sub Process';
-    const startEventName = 'Start Event Two';
-    const encodedConfig = JSON.stringify({
-      calledElement: 'Subprocess1-5',
-      processId: 5,
-      startEvent: 'node_10',
-      name: `${subProcessName}`,
-    }).replace(/"/g, '&#34;');
-
-    addNodeTypeToPaper(subProcessPosition, nodeTypes.task, 'switch-to-sub-process');
-    getElementAtPosition(subProcessPosition)
-      .click({ force: true });
-
-    cy.get('[data-test="inspector-container"]')
-      .contains('Process')
-      .next('.multiselect')
-      .click()
-      .find('.multiselect__content')
-      .contains(processName)
-      .click();
-
-    cy.get('[data-test="inspector-container"]')
-      .contains('Start Event')
-      .next('.multiselect')
-      .click()
-      .find('.multiselect__content')
-      .contains(startEventName)
-      .click();
-
-    assertDownloadedXmlContainsExpected(`<bpmn:callActivity id="node_3" name="${subProcessName}" calledElement="Subprocess1-5" pm:config="${encodedConfig}" />`);
   });
 
   it('Can switch task type when initially added', () => {
@@ -118,11 +82,5 @@ describe('Tasks', () => {
     modalCancel();
 
     getElementAtPosition(taskPosition).click().getType().should('equal', nodeTypes.subProcess);
-  });
-
-  it('Allows typing in name field on new sub process', () => {
-    addNodeTypeToPaper(taskPosition, nodeTypes.task, 'switch-to-sub-process');
-    typeIntoTextInput('[name=name]', testString);
-    cy.get('[name=name]').should('have.value', testString);
   });
 });
